backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.0.3
celery==5.3.4
redis==5.0.1
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1
numpy==1.24.3
Pillow==10.0.1
google-generativeai==0.3.2
requests==2.31.0
transformers==4.36.0
diffusers==0.24.0
accelerate==0.25.0
safetensors==0.4.1
imageio==2.31.6
imageio-ffmpeg==0.4.9
moviepy==1.0.8
segment-anything==1.0.0
midas==1.1.0
scipy==1.11.4
python-dotenv==1.0.0
